fix: General fixes and responsive adjustments on instance form
