+++ src/components/CenteredContainer/index.ts
@@ -1,0 +1,2 @@
+export { default } from './cmp'
+export type { CenteredContainerProps } from './types'
