+++ src/components/CenteredContainer/types.ts
@@ -1,0 +1,8 @@
+import { ReactNode } from 'react'
+
+export type CenteredContainerVariant = 'default' | 'dashboard'
+
+export type CenteredContainerProps = {
+  children: ReactNode
+  variant?: CenteredContainerVariant
+}
