+++ src/components/form/SelectInstanceImage/index.ts
@@ -1,1 +1,2 @@
 export { default } from './cmp'
+export type { SelectInstanceImageProps } from './types'
