+++ src/pages/dashboard/volume.tsx
@@ -1,20 +1,9 @@
-import BaseContainer from '@/components/Container'
 import HoldingRequirements from '@/components/HoldingRequirements'
 import NewVolume from '@/components/form/AddVolume'
 import { EntityType } from '@/helpers/constants'
 import { useNewVolumePage } from '@/hooks/pages/useNewVolumePage'
-import { Button, Col, Row, TextGradient } from '@aleph-front/aleph-core'
-import { ReactNode } from 'react'
-
-const Container = ({ children }: { children: ReactNode }) => (
-  <Row xs={1} lg={12} gap="0">
-    <Col xs={1} lg={10} lgOffset={2} xl={8} xlOffset={3} xxl={6} xxlOffset={4}>
-      <BaseContainer>
-        <div tw="max-w-[715px] mx-auto">{children}</div>
-      </BaseContainer>
-    </Col>
-  </Row>
-)
+import { Button, TextGradient } from '@aleph-front/aleph-core'
+import { default as Container } from '@/components/CenteredContainer'
 
 export default function NewVolumePage() {
   const {
